+++ core/scripts/updateFirmware.sh
@@ -1,10 +1,13 @@
 #! /bin/bash
-# PiZigate flash programmation script
-# updateFirmware.sh <action> <zigateport> <lib> [fwfile]
+# PiZigate & DIN flash programmation script
+# updateFirmware.sh <action> <zigatePort> <zigateType> <lib> [fwfile]
 #   where action = flash, check, eraseeeprom
-
-# ./updateFirmware.sh flash /dev/ttyS0 PiGpio ZiGate_v3.23-OPDM.bin
-# ./updateFirmware.sh eraseeeprom /dev/ttyS0 PiGpio ZiGate_v3.23-OPDM.bin
+#         zigateType = PI, PIv2, DIN
+#         lib = WiringPi, PiGpio
+#
+# Examples
+# ./updateFirmware.sh flash /dev/ttyS0 PI PiGpio ZiGate_v3.23-OPDM.bin
+# ./updateFirmware.sh eraseeeprom /dev/ttyS0 PI PiGpio ZiGate_v3.23-OPDM.bin
 
 # NOW=`date +"%Y-%m-%d %H:%M:%S"`
 # echo "[${NOW}] Démarrage de '$(basename $0)' $@"
@@ -13,23 +16,36 @@ echo "Démarrage de '$(basename $0)' $@"
 
 # Note: Startup directory is the one from the caller (ajax)
 #       It is then '/var/www/html/plugins/Abeille/core/ajax'
-PROG=${PWD}/../../tmp/JennicModuleProgrammer
+PROG_PIv1=${PWD}/../../tmp/JennicModuleProgrammer
 BUILD_DIR=${PWD}/../../resources/prog_jennic-0.7/build
+PROG_PIv2=${PWD}/../../resources/DK6Programmer
 FW_DIR=${PWD}/../../resources/fw_zigate
+DIN_SCRIPT=${PWD}/../scripts/flash_ZiGate-DIN.py
 
 # Checks
 echo "- Vérifications préliminaires"
 echo ${PWD}
 error=0
-if [ $# -lt 3 ]; then
+if [ $# -lt 4 ]; then
     echo "= ERREUR: Argument(s) manquant(s) !"
-    echo "=         updateFirmware.sh <action> <zigateport> <GpioLib> [fwfile]"
-    error=1
+    echo "=         updateFirmware.sh <action> <zigatePort> <zigateType> <GpioLib> [fwfile]"
+    exit 1
 fi
 ACTION=$1
 ZGPORT=$2
-GPIOLIB=$3
-FW=$4
+ZGTYPE=$3
+GPIOLIB=$4
+if [ ${ZGTYPE} != "PI" ] && [ ${ZGTYPE} != "PIv2" ] && [ ${ZGTYPE} != "DIN" ]; then
+    echo "= ERREUR: Type de Zigate invalide ! (PI, PIv2 ou DIN)"
+    echo "=         updateFirmware.sh <action> <zigatePort> <zigateType> <GpioLib> [fwfile]"
+    exit 1
+fi
+if [ ${ACTION} == "flash" ] && [ $# -lt 5 ]; then
+    echo "= ERREUR: Nom du FW manquant !"
+    echo "=         updateFirmware.sh <action> <zigatePort> <zigateType> <GpioLib> [fwfile]"
+    exit 1
+fi
+FW=$5
 
 if [ ${ACTION} != "flash" ] && [ ${ACTION} != "check" ] && [ ${ACTION} != "eraseeeprom" ]; then
     echo "= ERREUR: Action '${ACTION}' non supportée."
@@ -66,32 +82,54 @@ else
         echo "= ERREUR: Type de lib GPIO invalid (${GPIOLIB}) !"
         error=1
     fi
+fi
+if [ $error -ne 0 ]; then
+    exit 1
+fi
+echo "= Ok"
+
+if [ ${ZGTYPE} == "PI" ]; then
+    PROG=${PROG_PIv1}
+
     # (Re)Compiling Jennic programmer v0.7-Abeille
     echo "- Compilation du programmateur"
     pushd ${BUILD_DIR} >/dev/null
     sudo make
     if [ $? -ne 0 ]; then
         echo "= ERREUR: Compilation ratée !"
-        error=1
+        error=2
     fi
     # echo "= ERREUR: Programmateur Jennic manquant !"
-    # echo "=         ${PROG}"
+    # echo "=         ${PROG_PIv1}"
     # error=1
     popd >/dev/null
-    if [ ! -x ${PROG} ]; then
+    if [ ! -x ${PROG_PIv1} ]; then
         # Attempting to correct execution right
-        sudo chmod +x ${PROG} >/dev/null
+        sudo chmod +x ${PROG_PIv1} >/dev/null
     fi
-    if [ ! -x ${PROG} ]; then
+    if [ ! -x ${PROG_PIv1} ]; then
         echo "= ERREUR: Le programmateur Jennic n'est pas exécutable !"
-        echo "=         ${PROG}"
-        error=1
+        echo "=         ${PROG_PIv1}"
+        error=2
     fi
+elif [ ${ZGTYPE} == "PIv2" ]; then
+    PROG=${PROG_PIv2}
+elif [ ${ZGTYPE} == "DIN" ]; then
+    PROG=${DIN_SCRIPT}
+
+    command -v python3 >/dev/null
+    if [ $? -ne 0 ]; then
+        echo "= ERREUR: 'python3' manquant ou non exécutable !"
+        echo "=         Le package 'python3' est probablement mal installé."
+        error=2
+    fi
+else
+    echo "= ERREUR: Type ${ZGTYPE} inconnu !"
+    error=2
 fi
 if [ $error != 0 ]; then
-    exit 1
+    exit $error
 fi
-echo "= Ok"
 
 # Si check seulement on quitte ici
 if [ ${ACTION} == "check" ]; then
@@ -103,6 +141,7 @@ else
     echo "  Firmware: ${FW_PATH}"
 fi
 echo "  Port tty: ${ZGPORT}"
+echo "  Type    : ${ZGTYPE}"
 echo "  Lib GPIO: ${GPIOLIB}"
 echo "  Prog    : ${PROG}"
 
@@ -111,21 +150,21 @@ echo "  Prog    : ${PROG}"
 # port 2 = FLASH
 # Mode production: FLASH=1, RESET=0 puis 1
 # Mode flash: FLASH=0, RESET=0 puis 1
+if [ ${ZGTYPE} == "PI" ] || [ ${ZGTYPE} == "PIv2" ]; then
+    if [ ${GPIOLIB} == "WiringPi" ]; then
+        gpio mode 0 out
+        gpio mode 2 out
 
-
-if [ ${GPIOLIB} == "WiringPi" ]; then
-    gpio mode 0 out
-    gpio mode 2 out
-
-    # Passage en mode 'flash'
-    gpio write 2 0
-    sleep 1
-    gpio write 0 0
-    sleep 1
-    gpio write 0 1
-    sleep 1
-elif [ ${GPIOLIB} == "PiGpio" ]; then
-    python /var/www/html/plugins/Abeille/core/scripts/pizigateModeFlash.py
+        # Passage en mode 'flash'
+        gpio write 2 0
+        sleep 1
+        gpio write 0 0
+        sleep 1
+        gpio write 0 1
+        sleep 1
+    elif [ ${GPIOLIB} == "PiGpio" ]; then
+        python /var/www/html/plugins/Abeille/core/scripts/pizigateModeFlash.py
+    fi
 fi
 
 if [ ${ACTION} == "eraseeeprom" ]; then
@@ -140,7 +179,11 @@ if [ ${ACTION} == "eraseeeprom" ]; then
         status=0
     fi
 else
-    sudo ${PROG} -V 6 -P 115200 -v -f ${FW_PATH} -s ${ZGPORT} 2>&1
+    if [ ${ZGTYPE} == "DIN" ]; then
+        sudo python3 ${DIN_SCRIPT} -s ${ZGPORT} -b 250000 -f ${FW_PATH} 2>&1
+    else # PI or PIv2
+        sudo ${PROG} -V 6 -P 115200 -v -f ${FW_PATH} -s ${ZGPORT} 2>&1
+    fi
     if [ $? != 0 ]; then
         # echo "= ERREUR: Programmation impossible"
         status=2
@@ -152,14 +195,16 @@ else
 fi
 
 # Switch back to 'prod' mode
-if [ ${GPIOLIB} == "WiringPi" ]; then
-    gpio write 2 1
-    sleep 1
-    gpio write 0 0
-    sleep 1
-    gpio write 0 1
-elif [ ${GPIOLIB} == "PiGpio" ]; then
-    python /var/www/html/plugins/Abeille/core/scripts/resetPiZigate.py
+if [ ${ZGTYPE} == "PI" ] || [ ${ZGTYPE} == "PIv2" ]; then
+    if [ ${GPIOLIB} == "WiringPi" ]; then
+        gpio write 2 1
+        sleep 1
+        gpio write 0 0
+        sleep 1
+        gpio write 0 1
+    elif [ ${GPIOLIB} == "PiGpio" ]; then
+        python /var/www/html/plugins/Abeille/core/scripts/resetPiZigate.py
+    fi
 fi
 
 # Final status